move nodescanning mode to same binary as Scanner
