+++ cmd/clair/main.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	"github.com/stackrox/scanner/pkg/nodescanner"
 	"math/rand"
 	"net/http"
 	"net/http/pprof"
@@ -205,7 +206,14 @@ func main() {
 	// Parse command-line arguments
 	flag.CommandLine = flag.NewFlagSet(os.Args[0], flag.ExitOnError)
 	flagConfigPath := flag.String("config", "/etc/scanner/config.yaml", "Load configuration from the specified file.")
+	flagNodeScannerMode := flag.Bool("nodescanner", false, "Run Scanner binary in NodeScanning mode (this should only be used in Secured Clusters)")
 	flag.Parse()
+
+	if *flagNodeScannerMode {
+		log.Info("Starting Scanner in Node Scanning mode")
+		nodescanner.StartNodeScanningHTTPServer()
+		return
+	}
 
 	proxy.WatchProxyConfig(context.Background(), proxyConfigPath, proxyConfigFile, true)
