+++ image/scanner/scripts/nodescanning-entrypoint.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -euo pipefail
+
+exec /scanner -nodescanner
