+++ pkg/nodescanner/inventory/inventorizer.go
